Updates to human algorithm. Solves white cross.
Currently working on additional steps of the human algorithm.
Although it's not a mandatory requirement, I feel it will heavily familiarise me with every aspect of the cube - ready for more complex algorithms.
Added some boolean operators for when the program should call the solving function
Also when to hide the HUD - gives the program a cleaner look

Issues
- Going to disregard issues for higher cubes until I've finished the human algorithm catering for a 3x3x3
  (I'm still unsure the reason for random colour changes for bigger cubes as of yet.)
TODO:
- Add 2D visualisation of the cube
    - Allow user to custom scramble the cube by clicking 2D visualisation
- Research and discover methods of speeding up the programs performance as cube's above 25x25x25 puts FPS below 30.