Fixed 2D Visualisation.
Added a logbook PDF linked in the README - located in the repo.
Working on a searching algorithm
Have implemented a scoring function for each cubie all over the cube.
Uses a copy of a solved cube as reference to correctness of each cubie.
Algorithms are generated then applied according to which state of cube has the lowest score when each and every algorithm is applied to it.
Only caters for 3x3x3 for now.
Unknown issue regarding cube randomly changing scramble state...